Templating BatchFillerSparseTensor
